feat: atomic pattern 적용
